move genUrl.ts into /library and update imports
