+++ rundeckapp/grails-spa/packages/ui-trellis/src/library/services/jobEdit.ts
@@ -1,8 +1,8 @@
 import { JobOptionEdit, OptionValidation, SaveJobResponse } from "../types/jobs/JobEdit";
 import { api } from "./api";
-import { _genUrl } from "../../app/utilities/genUrl";
 import { getRundeckContext } from "../rundeckService";
 import { JobDefinition } from "../types/jobs/JobDefinition";
+import { _genUrl } from "../utilities/genUrl";
 
 const rundeckContext = getRundeckContext();
